Added get_games() call, broke api into seperate file
